Remove a harmful dtype argument

+++ chainer/functions/normalization/group_normalization.py
@@ -69,7 +69,7 @@ class GroupNormalization(function_node.FunctionNode):
         var += self.eps
         self.inv_std = var
         del var
-        xp.sqrt(self.inv_std, out=self.inv_std, dtype=x.dtype)
+        xp.sqrt(self.inv_std, out=self.inv_std)
         xp.reciprocal(self.inv_std, out=self.inv_std)
         x_hat *= self.inv_std[:, None]
